feat : 최종 게임 결과(수익)을 출력하는 기능 구현
1. BlackJackGame 클래스에 최종 게임 결과를 출력하는 메소드 구현.
2. Output 클래스의 수익을 출력하는 출력 메소드 자료형을 int -> double로 수정.

+++ src/main/java/domain/ui/Output.java
@@ -1,5 +1,5 @@
 /*
- * @(#)Output.java      0.7 2019.12.15
+ * @(#)Output.java      0.8 2019.12.15
  *
  * Copyright (c) 2019 lxxjn0
  */
@@ -10,7 +10,7 @@ package domain.ui;
  * 출력과 관련된 기능을 담당할 객체
  *
  * @author JUNYOUNG LEE (lxxjn0)
- * @version 0.7 2019.12.15
+ * @version 0.8 2019.12.15
  */
 public class Output {
     /**
@@ -224,7 +224,7 @@ public class Output {
      * @param userName User의 이름.
      * @param profit   해당 User의 최종 수익.
      */
-    public void printUserFinalProfit(String userName, int profit) {
+    public void printUserFinalProfit(String userName, double profit) {
         System.out.printf(USER_PROFIT_MESSAGE + NEW_LINE_STRING, userName, profit);
     }
